+++ NEWS.md
@@ -1,5 +1,11 @@
 # version 0.9-6
 
+* `st_sample` gains an argument, `by_polygon`, to more clevery sample `MULTIPOLYGON` geometries
+
+* `st_sample` accepts non-integer sample sizes, with a (suppressable) warning and handles values of size that would round to zero
+
+* `gdal_utils` adds utils `mdiminfo` and `mdimtranslate` (requires GDAL >= 3.1)
+
 * `st_union` gains an argument `is_coverage`, which, when set to `TRUE`, leads to much faster unioning in case features form a coverage (polygons don't overlap); #1462 by Don Baston
 
 * fix `gdal_utils("translate")` locking input file; #1452